Update tests for reduce_model
